--- frontend/src/Animal.tsx
@@ -1,4 +1,0 @@
-export type Animal = {
-    id: string,
-    name: string
-}
