+++ frontend/src/FormularContainer.tsx
@@ -1,6 +1,6 @@
-import {Button, Grid, Paper, styled, TextField, Typography} from "@mui/material";
+import {Button, Grid, TextField, Typography} from "@mui/material";
 import {useState} from "react";
-import {Animal} from "./Animal.tsx";
+import {Animal, Item} from "./Utils.tsx";
 import axios from "axios";
 
 
@@ -9,15 +9,6 @@ type Props = {
 
 }
 
-
-const Item = styled(Paper)(({theme}) => ({
-    backgroundColor: theme.palette.mode === 'dark' ? '#1A2027' : '#fff',
-    ...theme.typography.body2,
-    padding: theme.spacing(1),
-    textAlign: 'center',
-    color: theme.palette.text.secondary,
-    margin: 4
-}));
 
 function FormularContainer({setAnimals}: Props) {
