+++ frontend/src/Header.tsx
@@ -1,13 +1,5 @@
-import {Grid, Paper, styled, Typography} from "@mui/material";
-
-const Item = styled(Paper)(({theme}) => ({
-    backgroundColor: theme.palette.mode === 'dark' ? '#1A2027' : '#fff',
-    ...theme.typography.body2,
-    padding: theme.spacing(1),
-    textAlign: 'center',
-    color: theme.palette.text.secondary,
-    margin: 4
-}));
+import {Grid, Typography} from "@mui/material";
+import {Item} from "./Utils.tsx";
 
 
 function Header() {
