+++ frontend/src/Utils.tsx
@@ -1,0 +1,15 @@
+import {Paper, styled} from "@mui/material";
+
+export type Animal = {
+    id: string,
+    name: string
+}
+
+export const Item = styled(Paper)(({theme}) => ({
+    backgroundColor: theme.palette.mode === 'dark' ? '#1A2027' : '#fff',
+    ...theme.typography.body2,
+    padding: theme.spacing(1),
+    textAlign: 'center',
+    color: theme.palette.text.secondary,
+    margin: 4
+}));
